requirements.txt
Django==4.1.7
django-environ==0.10.0
uWSGI==2.0.21
psycopg2-binary==2.9.6